Booth/requirements.txt
pillow==10.2.0
opencv-python==4.5.3.56
python-escpos==3.1
numpy==1.26.4
Mastodon.py==1.8.1
configparser==6.0.1